backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "API for OpenDisco"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "fastapi[standard]>=0.120.4",
    "psycopg[binary]>=3.2.12",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.44",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
